skills/genai-maturity-assessor/requirements.txt
Jinja2>=3.1
PyYAML>=6.0
playwright>=1.40
pytest>=8.0
